Extracting s3 util functions to a designated file
