+++ nessie/lakectl_test.go
@@ -1,9 +1,6 @@
 package nessie
 
 import (
-	"bytes"
-	"net/http"
-	"os"
 	"strconv"
 	"testing"
 
@@ -352,60 +349,4 @@ func TestLakectlIngest(t *testing.T) {
 	RunCmdAndVerifySuccessWithFile(t, Lakectl()+" ingest --from s3://"+tempBucketName+"/ --to lakefs://"+repoName+"/"+mainBranch+"/", false, "lakectl_ingest", vars)
 	RunCmdAndVerifySuccessWithFile(t, Lakectl()+" fs ls lakefs://"+repoName+"/"+mainBranch+"/", false, "lakectl_fs_ls_after_ingest", vars)
 	RunCmdAndVerifySuccessWithFile(t, Lakectl()+" fs ls lakefs://"+repoName+"/"+mainBranch+"/ --recursive", false, "lakectl_fs_ls_after_ingest_recursive", vars)
-
-}
-
-func putS3Object(client *s3.S3, filePath, bucket, key string) error {
-	upFile, err := os.Open(filePath)
-	if err != nil {
-		return err
-	}
-	defer upFile.Close()
-
-	upFileInfo, _ := upFile.Stat()
-	var fileSize int64 = upFileInfo.Size()
-	fileBuffer := make([]byte, fileSize)
-	upFile.Read(fileBuffer)
-
-	// Put the file object to s3 with the file name
-	_, err = client.PutObject(&s3.PutObjectInput{
-		Bucket:               aws.String(bucket),
-		Key:                  aws.String(key),
-		ACL:                  aws.String("private"),
-		Body:                 bytes.NewReader(fileBuffer),
-		ContentLength:        aws.Int64(fileSize),
-		ContentType:          aws.String(http.DetectContentType(fileBuffer)),
-		ContentDisposition:   aws.String("attachment"),
-		ServerSideEncryption: aws.String("AES256"),
-	})
-
-	return err
-}
-
-func cleanS3Bucket(client *s3.S3, bucket string) error {
-	listObjectsInput := &s3.ListObjectsV2Input{Bucket: aws.String(bucket)}
-	for {
-		listObjectsOutput, err := client.ListObjectsV2(listObjectsInput)
-		if err != nil {
-			return err
-		}
-
-		for _, record := range listObjectsOutput.Contents {
-			_, err := client.DeleteObject(&s3.DeleteObjectInput{
-				Bucket: aws.String(bucket),
-				Key:    record.Key,
-			})
-			if err != nil {
-				return err
-			}
-		}
-
-		if !aws.BoolValue(listObjectsOutput.IsTruncated) {
-			break
-		}
-		listObjectsInput.ContinuationToken = listObjectsOutput.NextContinuationToken
-	}
-	_, err := client.DeleteBucket(&s3.DeleteBucketInput{Bucket: aws.String(bucket)})
-
-	return err
 }
